% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc1.R
\name{myfunc1}
\alias{myfunc1}
\title{Returns the value 1}
\usage{
myfunc1(...)
}
\arguments{
\item{...}{You can literally put anything in; it doesn't affect the output}
}
\value{
The integer 1
}
\description{
Returns the value 1
}
\examples{
myfunc1("thirteen")
}

